Add component.BuildODR and component.RegistryAccess
The optional interface BuildODR allows a plugin to define a separate
build function that will be called if the runner is operating in
ondemand runner (odr) mode. This allows plugins the ability to take
advantage of this unique run context.

The optional interface RegistryAccess allows registry's to simply
generate the access information for a OCI registry but not actually
push data to it. This allows a build plugin the ability to perform
the push, which allows plugins the ability to build and push all at once.