framework/resource: call destroy if SetState manually called
Fixes #37

This calls Destroy properly on a resource if SetState is manually
called. There is a big limitation here in that we don't know the exact
proper destroy order to call, but argmapper gets us most likely correct
due to state dependencies. And, as the comment in the code states, the
ordering doesn't matter for any practical use case we have today since
all manual SetStates today are single resources. Hopefully we switch
everything to resource manager soon so this doesn't matter.