New `StatusReport` function on resource manager
Introduces a new resource manager function `StatusReport`, that behind the scenes calls `StatusAll`, and then builds a status report
containing all the resources returned by the plugin's status functions. It also takes a stab at determining the overall health given
the list of resources.

If all resources return the same health, the overall health will be that health. If resources return different healths, the overall
health will be PARTIAL, and the health message will give more details.

Incidentally, this also lets the `Name`, `Type`, and `Platform` values of status report resources default to the parent resource's values
if they are unset. They generally will the same as the parent resource, so this is a nice convenience for plugin devs.